Move hooks closer to components
