+++ frontend/src/components/ArtworkFrameGlass/ArtworkFrameGlass.tsx
@@ -1,7 +1,7 @@
 import { StyledArtworkFrameGlass } from './styled';
 import {
   useArtworkFrameGlassParallax
-} from '../../services/use-artwork-frame-glass-parallax/use-artwork-frame-glass-parallax';
+} from './services/use-artwork-frame-glass-parallax/use-artwork-frame-glass-parallax';
 
 
 export const ArtworkFrameGlass = () => {
