+++ frontend/src/components/ArtworkFrameGlass/services/get-scrollable-parent/get-scrollable-parent.ts
@@ -1,0 +1,13 @@
+export const getScrollableParent = (node: HTMLElement | null): HTMLElement | null => {
+  if (!node) {
+    return null;
+  }
+
+  const overflowY = window.getComputedStyle(node).overflowY;
+
+  if (overflowY === 'auto' || overflowY === 'scroll') {
+    return node;
+  }
+
+  return getScrollableParent(node.parentElement);
+};
