+++ frontend/src/components/ArtworkList/services/use-on-scroll-to-end/use-on-scroll-to-end.ts
