+++ frontend/src/pages/MainPage/services/normalize-artwork-data/normalize-artwork-data.ts
